+++ packages/unity-bootstrap-theme/CHANGELOG.md
@@ -1,3 +1,57 @@
+# [@asu/unity-bootstrap-theme-v1.13.0](https://github.com/ASU/asu-unity-stack/compare/@asu/unity-bootstrap-theme-v1.12.0...@asu/unity-bootstrap-theme-v1.13.0) (2024-08-08)
+
+
+### Bug Fixes
+
+* **unity-bootstrap-theme:** added margin-bottom=0 on p ([a692f16](https://github.com/ASU/asu-unity-stack/commit/a692f161a1edf544708f47638c26850cc615d4f1))
+
+
+### Features
+
+* **unity-bootstrap-theme:** added css styles for p in donuts ([cc4a536](https://github.com/ASU/asu-unity-stack/commit/cc4a536486291725c67793c751170cffdd204c1e))
+* **unity-bootstrap-theme:** added new styles to the donut chart ([b3d10e7](https://github.com/ASU/asu-unity-stack/commit/b3d10e7b9b38ed6b3f0bd5aff18a9224cd8cd7ee))
+* **unity-bootstrap-theme:** update markup for charts and graphs ([86f7a04](https://github.com/ASU/asu-unity-stack/commit/86f7a041a7870c15c7062dc8fa11aa1f87911368))
+
+# [@asu/unity-bootstrap-theme-v1.12.0](https://github.com/ASU/asu-unity-stack/compare/@asu/unity-bootstrap-theme-v1.11.1...@asu/unity-bootstrap-theme-v1.12.0) (2024-08-07)
+
+
+### Bug Fixes
+
+* **unity-bootstrap-theme:** switch anchor to button on accordions header and added styles ([914e9ee](https://github.com/ASU/asu-unity-stack/commit/914e9ee5d57363284558f6f275fa7c6858a05b3b))
+
+
+### Features
+
+* **unity-bootstrap-theme:** added styles for focus on accordions ([145c4a1](https://github.com/ASU/asu-unity-stack/commit/145c4a10faf783a918973b7d300392d44d22d217))
+
+# [@asu/unity-bootstrap-theme-v1.11.1](https://github.com/ASU/asu-unity-stack/compare/@asu/unity-bootstrap-theme-v1.11.0...@asu/unity-bootstrap-theme-v1.11.1) (2024-08-05)
+
+
+### Bug Fixes
+
+* **unity-bootstrap-theme:** fix storybook image ([113039f](https://github.com/ASU/asu-unity-stack/commit/113039f558875094f0fbde25213dc3780173a189))
+
+# [@asu/unity-bootstrap-theme-v1.11.0](https://github.com/ASU/asu-unity-stack/compare/@asu/unity-bootstrap-theme-v1.10.1...@asu/unity-bootstrap-theme-v1.11.0) (2024-08-01)
+
+
+### Features
+
+* **unity-bootstrap-theme:** changed h4 to h2 and added styles for this change ([d1bd183](https://github.com/ASU/asu-unity-stack/commit/d1bd1830d88c3408a29a56c8caf9718c5df2af21))
+
+# [@asu/unity-bootstrap-theme-v1.10.1](https://github.com/ASU/asu-unity-stack/compare/@asu/unity-bootstrap-theme-v1.10.0...@asu/unity-bootstrap-theme-v1.10.1) (2024-07-31)
+
+
+### Bug Fixes
+
+* **unity-bootstrap-theme:** fix A11y issue with incorrect Hero Heading structure ([0b4c5e8](https://github.com/ASU/asu-unity-stack/commit/0b4c5e8efb92c7f9082f2e23f2659c6d3753f2d9))
+
+# [@asu/unity-bootstrap-theme-v1.10.0](https://github.com/ASU/asu-unity-stack/compare/@asu/unity-bootstrap-theme-v1.9.21...@asu/unity-bootstrap-theme-v1.10.0) (2024-07-31)
+
+
+### Features
+
+* **unity-bootstrap-theme:** modified _banner.scss, banners.templates.stories.js for accessibility ([044bab1](https://github.com/ASU/asu-unity-stack/commit/044bab18d2eefd925acf7f7479a4eec1819bdea5))
+
 # [@asu/unity-bootstrap-theme-v1.9.21](https://github.com/ASU/asu-unity-stack/compare/@asu/unity-bootstrap-theme-v1.9.20...@asu/unity-bootstrap-theme-v1.9.21) (2024-06-26)